refactored the Main content
- extracted the CoreConcept and created a CoreConceptSection component
- extracted the CoreExamples and created a ExamplesSection component
- created a Section component that utilized the newly created sections
- implemented the Forwarding props for the element attributes
