+++ src/components/CoreConcepts/CoreConcept.jsx
