+++ src/components/Section.jsx
@@ -1,0 +1,8 @@
+export default function Section({title, children, ...props}){
+    return (
+        <section {...props}>
+            <h2>{title}</h2>
+            {children}
+        </section>
+    );
+}
